requirements.txt
matplotlib==3.7.5
numpy
pandas==2.0.3
Pillow==11.2.1
seaborn==0.13.2
streamlit==1.40.1
streamlit_option_menu==0.4.0
